requirements.txt
streamlit>=1.35.0
plotly>=5.18.0
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
olca-ipc>=2.0.0
olca-schema>=0.0.12
